LayoutTests/storage/domstorage/localstorage/missing-arguments.html

<!DOCTYPE html>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
<script>

description('Ensures that we get exceptions when we do not include required arguments for localStorage methods.');

var notEnoughArguments = '"TypeError: Not enough arguments"';

shouldThrow('localStorage.key()', notEnoughArguments);
shouldThrow('localStorage.getItem()', notEnoughArguments);
shouldThrow('localStorage.setItem()', notEnoughArguments);
shouldThrow('localStorage.setItem("a")', notEnoughArguments);
shouldThrow('localStorage.removeItem()', notEnoughArguments);

</script>
<script src="../../../fast/js/resources/js-test-post.js"></script>
